Close #1089
Improve the import function to import *.tif and *.asc, to fix the issues after merging #1078:

When the dialog to select the pattern to recognize date and time, if a predefined pattern matches the filename, the pattern is already input when the dialog opens.
In the dialog to select file to import, the suffix of the files are not shown.
Improved the function related to coordinate system. When importing NetCDF data, user is asked to select the coordinate system of the project, first, and then asked to specify the coordinate system of data (if needed). The message is improved so that it is clear if user is asked to select the coordinate system of the project, or the data.